+++ .github/actions/sourceclear/entrypoint.sh
@@ -5,8 +5,7 @@ echo "Script saved to srcclr.sh"
 chmod a+x srcclr.sh
 echo "Script execution rights added"
 
-packages="cmf,cmf-cqrs,cmf-webpack-plugin,components,containers,datagrid,forms,icons,sagas,stepper,theme,dataviz,cmf-router,flow-designer"
-
+packages=$(yarn --silent workspaces info  | jq '.[].location' | sed 's/\"//g')
 # set comma as internal field separator for the string list
 Field_Separator=$IFS
 IFS=,
@@ -18,9 +17,10 @@ echo "Scan completed on . (root)"
 
 for folder in $packages;
 do
-    echo "Starting scan on ./packages/$folder ..."
-    ./srcclr.sh scan ./packages/$folder
-    echo "Scan completed on ./packages/$folder"
+    echo "Starting scan on ./$folder ..."
+    ln -s yarn.lock ./$folder/yarn.lock
+    ./srcclr.sh scan ./$folder
+    echo "Scan completed on ./$folder"
 done
 
 # set back default field separator