+++ packages/components/src/Datalist/Datalist.component.js
@@ -67,7 +67,7 @@ function findEntry(titleMap, attributeName, attributeValue = '') {
 			}
 		} else if (
 			!entry.disabled &&
-			entry[attributeName].toLowerCase() === attributeValue.toLowerCase()
+			String(entry[attributeName]).toLowerCase() === String(attributeValue).toLowerCase()
 		) {
 			// entry is {name, value}
 			return entry;
@@ -78,6 +78,10 @@ function findEntry(titleMap, attributeName, attributeValue = '') {
 }
 
 function getEntryFromName(titleMap, name, restricted) {
+	if (name === '') {
+		return { name, value: '' };
+	}
+
 	const entry = findEntry(titleMap, 'name', name);
 	if (entry) {
 		return entry;
@@ -147,7 +151,7 @@ function Datalist(props) {
 			setFilterValue(entry.name);
 		}
 		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [props.value]);
+	}, [props.value, props.titleMap]);
 
 	// Suggestion display syntaxic sugar
 	const resetSelection = () =>