+++ tools/scripts-core/scripts/test-ng.js
@@ -5,7 +5,7 @@ const { getPreset } = require('../utils/preset');
 const karma = resolveBin('karma');
 
 module.exports = function test(env, presetApi, options) {
-	const presetName = presetApi.getUserConfig(['preset'], 'talend');
+	const presetName = presetApi.getUserConfig(['preset'], '@talend/scripts-preset-react-lib');
 	const preset = getPreset(presetName);
 	const karmaConfigPath = preset.getKarmaConfigurationPath(presetApi);