+++ packages/playground/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @talend/ui-playground
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [9f84c4ad7]
+- Updated dependencies [65599e1a5]
+- Updated dependencies [20e907887]
+- Updated dependencies [e04e3910f]
+- Updated dependencies [e04e3910f]
+  - @talend/react-datagrid@9.0.0
+  - @talend/react-dataviz@2.1.1
+  - @talend/react-components@6.46.2
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [d1815c0af]
+- Updated dependencies [4605d2ce4]
+- Updated dependencies [d1815c0af]
+- Updated dependencies [d1815c0af]
+  - @talend/react-datagrid@8.0.0
+  - @talend/react-forms@7.1.0
+  - @talend/react-dataviz@2.1.0
+
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [9f3cad9bd]
+  - @talend/react-dataviz@2.0.0
+
 ## 1.0.2
 
 ### Patch Changes